tests/requirements.txt
requests==2.22.0
pytest==5.2.2
pyyaml==5.1.2